refactor: 整体代码的调整以及重构

+++ recommend-server/src/main/java/com/recommend/service/feature/FeatureExtractor.java
@@ -2,67 +2,68 @@ package com.recommend.service.feature;
 
 import com.recommend.common.entity.GameMaster;
 import com.recommend.common.entity.UserProfile;
-import org.springframework.stereotype.Component;
 import lombok.extern.slf4j.Slf4j;
-import java.util.Map;
+import org.springframework.stereotype.Component;
+
 import java.util.HashMap;
+import java.util.Map;
 
 @Component
 @Slf4j
 public class FeatureExtractor {
-    
+
     /**
      * 提取排序特征
      */
     public RankingFeatures extractFeatures(GameMaster master, UserProfile userProfile) {
         return RankingFeatures.builder()
-            .master(master)
-            .relevanceScore(calculateRelevanceScore(master, userProfile))
-            .tagSimilarity(calculateTagSimilarity(master, userProfile))
-            .regionMatch(calculateRegionMatch(master, userProfile))
-            .ageMatch(calculateAgeMatch(master, userProfile))
-            .gameTypeMatch(calculateGameTypeMatch(master, userProfile))
-            .rating(master.getScore() != null ? master.getScore().doubleValue() : 0.0)
-            .orderCount(master.getOrderCount())
-            .completionRate(calculateCompletionRate(master))
-            .responseTime(calculateResponseTime(master))
-            .diversityScore(calculateDiversityScore(master))
-            .categoryCount(calculateCategoryCount(master))
-            .noveltyScore(calculateNoveltyScore(master))
-            .popularityScore(calculatePopularityScore(master))
-            .viewCount(calculateViewCount(master))
-            .likeCount(calculateLikeCount(master))
-            .commentCount(calculateCommentCount(master))
-            .priceScore(calculatePriceScore(master))
-            .priceRatio(calculatePriceRatio(master))
-            .isPriceReasonable(isPriceReasonable(master))
-            .timeDecayScore(calculateTimeDecayScore(master))
-            .isOnline(isOnline(master))
-            .availabilityScore(calculateAvailabilityScore(master))
-            .personalizedScore(calculatePersonalizedScore(master, userProfile))
-            .historicalPreference(calculateHistoricalPreference(master, userProfile))
-            .behaviorSimilarity(calculateBehaviorSimilarity(master, userProfile))
-            .customFeatures(extractCustomFeatures(master, userProfile))
-            .build();
+                .master(master)
+                .relevanceScore(calculateRelevanceScore(master, userProfile))
+                .tagSimilarity(calculateTagSimilarity(master, userProfile))
+                .regionMatch(calculateRegionMatch(master, userProfile))
+                .ageMatch(calculateAgeMatch(master, userProfile))
+                .gameTypeMatch(calculateGameTypeMatch(master, userProfile))
+                .rating(master.getScore() != null ? master.getScore().doubleValue() : 0.0)
+                .orderCount(master.getOrderCount())
+                .completionRate(calculateCompletionRate(master))
+                .responseTime(calculateResponseTime(master))
+                .diversityScore(calculateDiversityScore(master))
+                .categoryCount(calculateCategoryCount(master))
+                .noveltyScore(calculateNoveltyScore(master))
+                .popularityScore(calculatePopularityScore(master))
+                .viewCount(calculateViewCount(master))
+                .likeCount(calculateLikeCount(master))
+                .commentCount(calculateCommentCount(master))
+                .priceScore(calculatePriceScore(master))
+                .priceRatio(calculatePriceRatio(master))
+                .isPriceReasonable(isPriceReasonable(master))
+                .timeDecayScore(calculateTimeDecayScore(master))
+                .isOnline(isOnline(master))
+                .availabilityScore(calculateAvailabilityScore(master))
+                .personalizedScore(calculatePersonalizedScore(master, userProfile))
+                .historicalPreference(calculateHistoricalPreference(master, userProfile))
+                .behaviorSimilarity(calculateBehaviorSimilarity(master, userProfile))
+                .customFeatures(extractCustomFeatures(master, userProfile))
+                .build();
     }
-    
+
     private Double calculateRelevanceScore(GameMaster master, UserProfile userProfile) {
         // TODO: 实现相关性得分计算
         return 0.5;
     }
-    
+
     private Double calculateTagSimilarity(GameMaster master, UserProfile userProfile) {
         // TODO: 实现标签相似度计算
         return 0.5;
     }
-    
+
     private Double calculateRegionMatch(GameMaster master, UserProfile userProfile) {
         if (master.getRegion() != null && userProfile.getRegion() != null) {
             return master.getRegion().equals(userProfile.getRegion()) ? 1.0 : 0.0;
         }
         return 0.5;
     }
-    
+
     private Double calculateAgeMatch(GameMaster master, UserProfile userProfile) {
         if (master.getAge() != null && userProfile.getAge() != null) {
             int ageDiff = Math.abs(master.getAge() - userProfile.getAge());
@@ -70,101 +71,101 @@ public class FeatureExtractor {
         }
         return 0.5;
     }
-    
+
     private Double calculateGameTypeMatch(GameMaster master, UserProfile userProfile) {
         // TODO: 实现游戏类型匹配计算
         return 0.5;
     }
-    
+
     private Double calculateCompletionRate(GameMaster master) {
         // TODO: 实现完成率计算
         return 0.9;
     }
-    
+
     private Double calculateResponseTime(GameMaster master) {
         // TODO: 实现响应时间计算
         return 0.8;
     }
-    
+
     private Double calculateDiversityScore(GameMaster master) {
         // TODO: 实现多样性得分计算
         return 0.6;
     }
-    
+
     private Integer calculateCategoryCount(GameMaster master) {
         // TODO: 实现类别数量计算
         return 3;
     }
-    
+
     private Double calculateNoveltyScore(GameMaster master) {
         // TODO: 实现新颖度得分计算
         return 0.4;
     }
-    
+
     private Double calculatePopularityScore(GameMaster master) {
         // TODO: 实现热门度得分计算
         return 0.7;
     }
-    
-    private Integer calculateViewCount(GameMaster master) {
+
+    private Long calculateViewCount(GameMaster master) {
         // TODO: 实现观看数计算
-        return 1000;
+        return 1000L;
     }
-    
-    private Integer calculateLikeCount(GameMaster master) {
+
+    private Long calculateLikeCount(GameMaster master) {
         // TODO: 实现点赞数计算
-        return 100;
+        return 100L;
     }
-    
-    private Integer calculateCommentCount(GameMaster master) {
+
+    private Long calculateCommentCount(GameMaster master) {
         // TODO: 实现评论数计算
-        return 50;
+        return 50L;
     }
-    
+
     private Double calculatePriceScore(GameMaster master) {
         // TODO: 实现价格得分计算
         return 0.6;
     }
-    
+
     private Double calculatePriceRatio(GameMaster master) {
         // TODO: 实现价格比率计算
         return 0.8;
     }
-    
+
     private Boolean isPriceReasonable(GameMaster master) {
         // TODO: 实现价格合理性判断
         return true;
     }
-    
+
     private Double calculateTimeDecayScore(GameMaster master) {
         // TODO: 实现时间衰减得分计算
         return 0.9;
     }
-    
+
     private Boolean isOnline(GameMaster master) {
         return master.getStatus() != null && master.getStatus() == 1;
     }
-    
+
     private Double calculateAvailabilityScore(GameMaster master) {
         // TODO: 实现可用性得分计算
         return 0.8;
     }
-    
+
     private Double calculatePersonalizedScore(GameMaster master, UserProfile userProfile) {
         // TODO: 实现个性化得分计算
         return 0.7;
     }
-    
+
     private Double calculateHistoricalPreference(GameMaster master, UserProfile userProfile) {
         // TODO: 实现历史偏好计算
         return 0.6;
     }
-    
+
     private Double calculateBehaviorSimilarity(GameMaster master, UserProfile userProfile) {
         // TODO: 实现行为相似度计算
         return 0.5;
     }
-    
+
     private Map<String, Double> extractCustomFeatures(GameMaster master, UserProfile userProfile) {
         Map<String, Double> features = new HashMap<>();
         // TODO: 实现自定义特征提取
